Refactor VoteMember and CustomSwitch components; move to extra directory and update imports
