+++ components/extra/FrequencyBar.tsx
