+++ components/extra/VerificationBar.tsx
